leaderboard controller, service and repository

+++ questevent-backend/src/main/java/com/questevent/dto/LeaderboardDTO.java
@@ -4,6 +4,6 @@ public record LeaderboardDTO(
         Long userId,
         String userName,
         Long participationCount,
-        Long gems,
+        Integer gems,
         Double score
 ) {}
